Added unit tests for the search endpoint. Fixes for endpoints and CRUD operations. Added config.py for loading env variables. Minor fixes after testing. Updated README.

+++ app/api/routers/products.py
@@ -1,5 +1,5 @@
 from typing import Annotated
-from fastapi import APIRouter, Depends, Query
+from fastapi import APIRouter, Depends, HTTPException, Query, Response
 from sqlalchemy.orm import Session
 from app.api.deps import get_db
 from app.crud.products import create_product_crud, delete_product_crud, list_all_products, list_specific_product, search_for_product, update_product_crud
@@ -20,16 +20,26 @@ def search_products(query: Annotated[ProductParams, Query()], db: Session = Depe
 
 @router.get("/{product_id}", response_model=ProductRead)
 def get_product_by_id(product_id: int, db: Session = Depends(get_db)):
-    return list_specific_product(db, product_id)
+    product = list_specific_product(db, product_id)
+    if product is None:
+        raise HTTPException(status_code=404, detail="Product not found")
+    return product
 
-@router.post("/", response_model=ProductRead)
+@router.post("/", response_model=ProductRead, status_code=201)
 def create_product(product: ProductCreate, db: Session = Depends(get_db)):
     return create_product_crud(db, product)
 
 @router.patch("/{product_id}", response_model=ProductRead)
-def update_product_by_id(product_id: int, category: ProductUpdate, db: Session = Depends(get_db)):
-    return update_product_crud(db, product_id, category)
+def update_product_by_id(product_id: int, product: ProductUpdate, db: Session = Depends(get_db)):
+    updated_product = update_product_crud(db, product_id, product)
+    if updated_product is None:
+        raise HTTPException(status_code=404, detail="Product not found")
+    return updated_product
 
-@router.delete("/{product_id}")
+@router.delete("/{product_id}", status_code=204)
 def delete_product_by_id(product_id: int, db: Session = Depends(get_db)):
+    product = list_specific_product(db, product_id)
+    if product is None:
+        raise HTTPException(status_code=404, detail="Product not found")
     delete_product_crud(db, product_id)
+    return Response(status_code=204)
